Feedback from the previous PR: https://github.com/Microsoft/gulp-core-build-serve/pull/9
